Create initialized sites subdirectory
Summary:
Database files for initialized sites should be store in an existing
directory, service create only the database directory itself.

The content of the `ensure_parent_directory_exists` has been wrote
by ChatGPT, as part of an experiment to evaluate suitability to help
to write, review bugs and development. This is a trivial case where
it succeeds to write the body of a method given its signature.

Test Plan:
  - Delete initialized, run `alkane deploy`
  - Run it again
  - Create initialized as a file

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D3016